refactor(inference): drop underscore prefixes from module filenames
Rename _stubs.py -> predictor.py, _app.py -> app.py, _cam.py -> cam.py,
_cli.py -> cli.py, _optional.py -> optional.py to match the naming
convention used across radiologist-core, radiologist-etl, and
radiologist-utils (no underscore prefix on internal implementation files).

Update all internal imports inside the renamed modules and update every
test that referenced the old underscore-prefixed module paths.
