+++ radiologist-inference/tests/test_cli.py
@@ -35,7 +35,7 @@ def _get_runner_and_app():
     """Import typer testing utilities and the CLI app."""
     from typer.testing import CliRunner
 
-    from radiologist.inference._cli import app
+    from radiologist.inference.cli import app
 
     return CliRunner(), app
 
@@ -56,7 +56,7 @@ def test_predict_exits_0_on_valid_image_and_model(tmp_path):
         predicted_class="NORMAL",
     )
 
-    with patch("radiologist.inference._cli.Predictor") as MockPredictor:
+    with patch("radiologist.inference.cli.Predictor") as MockPredictor:
         instance = MagicMock()
         instance.predict.return_value = mock_prediction
         MockPredictor.from_path.return_value = instance
@@ -74,7 +74,7 @@ def test_pull_exits_0_on_valid_artifact(tmp_path):
     """pull command downloads model and exits 0 when artifact is retrievable."""
     runner, app = _get_runner_and_app()
 
-    with patch("radiologist.inference._cli.pull_model") as mock_pull:
+    with patch("radiologist.inference.cli.pull_model") as mock_pull:
         mock_pull.return_value = str(tmp_path / "model.onnx")
 
         result = runner.invoke(
@@ -93,7 +93,7 @@ def test_predict_exits_1_on_unreadable_image(tmp_path):
     fake_model = tmp_path / "model.onnx"
     fake_model.write_bytes(b"FAKE")
 
-    with patch("radiologist.inference._cli.Predictor") as MockPredictor:
+    with patch("radiologist.inference.cli.Predictor") as MockPredictor:
         MockPredictor.from_path.side_effect = Exception("cannot load model")
 
         result = runner.invoke(
@@ -111,7 +111,7 @@ def test_predict_exits_1_on_unreadable_model(tmp_path):
     fake_image = tmp_path / "chest.jpg"
     fake_image.write_bytes(b"FAKE")
 
-    with patch("radiologist.inference._cli.Predictor") as MockPredictor:
+    with patch("radiologist.inference.cli.Predictor") as MockPredictor:
         MockPredictor.from_path.side_effect = FileNotFoundError("model not found")
 
         result = runner.invoke(
@@ -126,7 +126,7 @@ def test_pull_exits_1_on_unretrievable_artifact(tmp_path):
     """pull command exits 1 when artifact cannot be retrieved."""
     runner, app = _get_runner_and_app()
 
-    with patch("radiologist.inference._cli.pull_model") as mock_pull:
+    with patch("radiologist.inference.cli.pull_model") as mock_pull:
         mock_pull.side_effect = RuntimeError("W&B download failed")
 
         result = runner.invoke(
@@ -139,7 +139,7 @@ def test_pull_exits_1_on_unretrievable_artifact(tmp_path):
 
 def test_cli_entry_point_raises_runtime_error_when_typer_absent():
     """Invoking cli entry point without typer raises RuntimeError naming 'cli' extra."""
-    import radiologist.inference._cli as cli_mod
+    import radiologist.inference.cli as cli_mod
 
     with patch.object(cli_mod, "_typer", None):
         with pytest.raises(RuntimeError, match="cli"):
